Move export to frk command from model plot to model.
