+++ ApplicationCode/Commands/RicExportFractureModelToFileFeature.h
@@ -1,6 +1,6 @@
 /////////////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (C) 2020-    Equinor ASA
+//  Copyright (C) 2020-     Equinor ASA
 //
 //  ResInsight is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -18,17 +18,18 @@
 
 #pragma once
 
-#include <vector>
-
-class RimFractureModel;
-class QString;
-class QTextStream;
+#include "cafCmdFeature.h"
 
 //==================================================================================================
-//
+///
 //==================================================================================================
-class RifFractureModelPlotExporter
+class RicExportFractureModelToFileFeature : public caf::CmdFeature
 {
-public:
-    static bool writeToDirectory( RimFractureModel* fractureModel, bool useDetailedFluidLoss, const QString& directoryPath );
+    CAF_CMD_HEADER_INIT;
+
+protected:
+    // Overrides
+    bool isCommandEnabled() override;
+    void onActionTriggered( bool isChecked ) override;
+    void setupActionLook( QAction* actionToSetup ) override;
 };
