+++ ApplicationCode/FileInterface/CMakeLists_files.cmake
@@ -54,7 +54,7 @@ ${CMAKE_CURRENT_LIST_DIR}/RifSurfaceImporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifRoffReader.h
 ${CMAKE_CURRENT_LIST_DIR}/RifColorLegendData.h
 ${CMAKE_CURRENT_LIST_DIR}/RifElasticPropertiesReader.h
-${CMAKE_CURRENT_LIST_DIR}/RifFractureModelPlotExporter.h
+${CMAKE_CURRENT_LIST_DIR}/RifFractureModelExporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelGeologicalFrkExporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelDeviationFrkExporter.h
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelPerfsFrkExporter.h
@@ -119,7 +119,7 @@ ${CMAKE_CURRENT_LIST_DIR}/RifSurfaceImporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifRoffReader.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifColorLegendData.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifElasticPropertiesReader.cpp
-${CMAKE_CURRENT_LIST_DIR}/RifFractureModelPlotExporter.cpp
+${CMAKE_CURRENT_LIST_DIR}/RifFractureModelExporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelGeologicalFrkExporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelDeviationFrkExporter.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RifFractureModelPerfsFrkExporter.cpp
