+++ ApplicationCode/FileInterface/RifFractureModelExporter.h
@@ -1,6 +1,6 @@
 /////////////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (C) 2020-     Equinor ASA
+//  Copyright (C) 2020-    Equinor ASA
 //
 //  ResInsight is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -18,18 +18,17 @@
 
 #pragma once
 
-#include "cafCmdFeature.h"
+#include <vector>
+
+class RimFractureModel;
+class QString;
+class QTextStream;
 
 //==================================================================================================
-///
+//
 //==================================================================================================
-class RicExportFractureModelPlotToFileFeature : public caf::CmdFeature
+class RifFractureModelExporter
 {
-    CAF_CMD_HEADER_INIT;
-
-protected:
-    // Overrides
-    bool isCommandEnabled() override;
-    void onActionTriggered( bool isChecked ) override;
-    void setupActionLook( QAction* actionToSetup ) override;
+public:
+    static bool writeToDirectory( RimFractureModel* fractureModel, bool useDetailedFluidLoss, const QString& directoryPath );
 };
